Add the client to the repo
I was previously going to split up the client and the server into
separate repos, however, that does not seem to be woth it given the fact
that the app will probably be living on Heroku most of the time and it
makes deployment more difficult.  I'm still going to be treating them as
essentially separate projects, however, there is a small chance I'll
give in and share code at some point as well.  We'll see.