Fix issues with logging out (and other minor optimizations)
Logging out was encountering some issues with the current view still
listening and that generating errors.  This adds immediate unbinding of
the current view.

At this point, it is pretty obvious where we could just wrap all the
route functions, so let's do that at some point.

This also bumps the version to 0.1.1