Restructure of how argument validation is done
The previous route for argument validation was that we would try the
validator function (or magic argument) and would then, if the output of
that was null, check to see if there was a default.  This let to
multiple errors with functions throwing themselves on invalid values
(such as null) as well as default values being set when the user had
sent an invalid argument if the function did not throw an error.

With this revamp, the api handler will first check if no value was set
and, if that was the case, will check to see if either the argument was
opional or if there is a default value.  If a value was set, it will
then be processed and a null return from the handler will be interpreted
as an indication that the value passed in by the user should be treated
as null.

This also includes a change to how the validator file is set up in that
every property in it is now a function which returns the actual
validator function to use.  Most of them take no arguments but are still
functions for consistency.  The magic object types have also been
removed.