Better define the lifetime of a form
The previous way of looking at forms, as being either active or not, did
not properly capture the idea that a form could be released to clubs,
and then retired.  To solve this, forms now have the properties of
active and published with a form being available to start only when it
is both active and published.