Add a basic view for creating a form
This adds a very rudimentary view for creating a new form, and adds a
link to it to the district management page.

Right now, it will work for fully setting up a form with the exception
of table and select type questions, which is actually pretty good I
think.  Those will be added soon.