Add the ability to fully define select elements in the form editor
This allows a user to specify the value for select options.  Because of
the way properties are handled for forms, this does behave slightly
hackily.  We make a large list of input items and, whenever any of them
are changed, we just completely rebuild the array using whenever is
well-defined.