Add better support for sourcemaps
So, I'm not thrilled about either of these changes, and they both feel
really hacky.  First of all, this causes the site to serve the actual
source code for the client  if requested.  This also runs sed on all
output files and removes an annoying line added by uglifyjs that has a
bad path for the source map.  Between these two changes, sentry source
mapping actually works.  But it sucks :(.