Reduce number of queries to open the district landing page
One of the things that has always annoyed me about how I set this up is
that loading the district landing page requires a separate query to be
done for each club to get its reports.  With this change, we now just do
one query to get reports for all clubs in the district and then get the
reports we need out of that for the club landing page.