Add new fields to report schema to identify model it is for
Previously, reports were referred to by modelType and idModel.  This
worked fine, and allowed all information to be accessible, and did a
great job of reducing redundancy, however, it unfortunately resulted in
really annoying queries when trying to get statistics for an entire
district.

To get around this, we are now storing idClub and idDistrict in addition
to idModel.  These are self-explanatory in name.

idModel will most likely be removed in a patch very shortly.